This commit was generated by cvs2svn to compensate for changes in r11869, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@11870 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/6510core.c
@@ -163,7 +163,7 @@
 
 /* Stack operations. */
 
-#define PUSH(val) ((PAGE_ONE)[(reg_sp--)] = (val))
+#define PUSH(val) ((PAGE_ONE)[(reg_sp--)] = ((BYTE)(val)))
 #define PULL()    ((PAGE_ONE)[(++reg_sp)])
 
 #ifdef DEBUG
